% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/importar_datos.R
\name{mindicador_importar_datos}
\alias{mindicador_importar_datos}
\title{Obtiene datos de mindicador.cl}
\usage{
mindicador_importar_datos(
  series = "uf",
  anios = 2020,
  tipo = "data.frame",
  max_intentos = 5,
  usar_cache = F,
  archivo = NULL
)
}
\arguments{
\item{series}{Cualquiera de los codigos de \code{mindicador::indicadores}}

\item{anios}{foo bar}

\item{tipo}{Por ahora solo "data.frame"}

\item{max_intentos}{Por defecto son cinco}

\item{usar_cache}{Valor logico para leer y escribir en cache. Si es \code{TRUE}, los resultados quedan cacheados en
memoria si \code{archivo} es \code{NULL} o en disco si \code{archivo} es una cadena de texto.}

\item{archivo}{Ruta del archivo para leer y escribir la cache.}
}
\description{
Se conecta a la API y realiza algunas transformaciones al input en formato JSON
para facilitar el uso de estos datos con distintas funciones de R
}
\examples{
\dontrun{
# foo bar
}
}
\keyword{functions}
